Added vimspector
Configured vimspector and added some custom mappings, I'm using
`vim-maximizer` to be able to jump to any of the debugger's panes
and occupy the whole screen.

The whole configuration was extrated from the following video:
    - https://www.youtube.com/watch?v=AnTX2mtOl9Q

Vimspector works using adapters, which you install throug
`:VimspectorInstal`. In the case of python I'm using the `debugpy`
adapter.
Another important thing to note is that you need to configure the run
in a `.vimspector.json` file in your project.

Closes: #33